==> susquehanna/notebooks_100k/10. Performance metrics table.ipynb <==
import sys
import pandas as pd

sys.path.append("..")
import os
import rbf_functions

rbfs = [
    rbf_functions.original_rbf,
    rbf_functions.squared_exponential_rbf,
    rbf_functions.inverse_quadratic_rbf,
    rbf_functions.inverse_multiquadric_rbf,
    rbf_functions.exponential_rbf,
    rbf_functions.matern32_rbf,
    rbf_functions.matern52_rbf,
]

# Load data

gddata = {}
eidata = {}
hvdata = {}
varlist = ["10", "20", "30", "40", "50", "60", "70", "80", "90", "100"]
for entry in rbfs:
    rbf = entry.__name__
    df_gd = pd.read_csv(f"gd_all/{rbf}_gd_all.csv")
    df_ei = pd.read_csv(f"ei_all/{rbf}_ei_all.csv")
    df_hv = pd.read_csv(f"hv_allref/{rbf}_all_ref_hv.csv")
    varlist = ["10", "20", "30", "40", "50", "60", "70", "80", "90", "100"]
    gddata[rbf] = {}
    eidata[rbf] = {}
    hvdata[rbf] = {}
    for i in varlist:  # add metric value at 100.000 nfe to dict for each seed
        gddata[rbf][i] = df_gd[i].dropna().iloc[-1]
        eidata[rbf][i] = df_ei[i].dropna().iloc[-1]
        hvdata[rbf][i] = df_hv[i].dropna().iloc[-1]
df_gd_m = pd.DataFrame.from_dict(gddata, orient="index")
df_ei_m = pd.DataFrame.from_dict(eidata, orient="index")
df_hv_m = pd.DataFrame.from_dict(hvdata, orient="index")

paretosets = {}
for entry in rbfs:
    name = entry.__name__
    # load variables
    varlist = []
    variables = []

    output_dir = os.path.abspath("../notebooks/refsets/")
    for filename in os.listdir(output_dir):
        if filename == f"{name}_refset.csv":
            varlist.append(filename[:-4])
            df_vars = pd.read_csv(f"{output_dir}\{filename}")
            paretosets[name] = len(df_vars)
    variables = df_vars.values
#     print(f"Loaded: {', '.join(varlist)}")

paretosets

df_sol = pd.DataFrame.from_dict(paretosets, orient="index")
df_sol

df_table = pd.DataFrame()
df_table["solutions"] = df_sol[0]
df_table["Generational Distance avg"] = df_gd_m.T.mean()
df_table["Epsilon Indicator avg"] = df_ei_m.T.mean()
df_table["Hypervolume avg"] = df_hv_m.T.mean()
df_table
