==> susquehanna/notebooks_100k/13. Kruskal-Wallis.ipynb <==
# When comparing two data sets using statistical tools, it is not
# sufficient to simply compare their average or median values. This is because
# such results can be skewed by randomness. For example, suppose we are
# calculating the median values of ten seeds. If one algorithm gets "lucky"
# and happens to use more above-average seeds, the estimated median will
# be skewed. Therefore, it is necessary to check the statistical significance of
# results. This is exactly what the indifferent entries are displaying. To determine
# statistical significance, the Kruskal-Wallis and Mann-Whitney U tests with 95% confidence
# intervals can be used.

# The Kruskal-Wallis H-test tests the null hypothesis that the population median of all
# of the groups are equal. It is a non-parametric version of ANOVA.
# The test works on 2 or more independent samples, which may have different sizes.
# Note that rejecting the null hypothesis does not indicate which of the groups differs.
# Post hoc comparisons between groups are required to determine which groups are different.

import numpy as np
from scipy import stats
import sys
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt

plt.rcParams["figure.figsize"] = [12, 8]
sys.path.append("..")
import rbf_functions

rbfs = [
    rbf_functions.original_rbf,
    rbf_functions.squared_exponential_rbf,
    rbf_functions.inverse_quadratic_rbf,
    rbf_functions.inverse_multiquadric_rbf,
    rbf_functions.exponential_rbf,
    rbf_functions.matern32_rbf,
    rbf_functions.matern52_rbf,
]

# Kruskal-Wallis HV local reference set

gddata = {}
eidata = {}
hvdata = {}
varlist = ["10", "20", "30", "40", "50", "60", "70", "80", "90", "100"]
for entry in rbfs:
    rbf = entry.__name__
    #     df_gd_data = pd.read_csv(f"gd_all/{rbf}_gd_all.csv")
    #     df_ei_data = pd.read_csv(f"ei_all/{rbf}_ei_all.csv")
    df_hv_data = pd.read_csv(f"hv/{rbf}_hv.csv")
    gddata[rbf] = {}
    eidata[rbf] = {}
    hvdata[rbf] = {}
    for i in varlist:  # add max hv value at 100.000 nfe to dict for each seed
        #         gddata[rbf][i] = df_gd_data[i].dropna().iloc[-1]
        #         eidata[rbf][i] = df_ei_data[i].dropna().iloc[-1]
        hvdata[rbf][i] = df_hv_data[i].dropna().iloc[-1]
df_gd = pd.DataFrame.from_dict(gddata, orient="index")
df_ei = pd.DataFrame.from_dict(eidata, orient="index")
df_hv = pd.DataFrame.from_dict(hvdata, orient="index")

df_kruskal = pd.DataFrame(columns=df_hv.index)
df_pval = pd.DataFrame(columns=df_hv.index)
for i, row in df_hv.iterrows():
    for j, row2 in df_hv.iterrows():
        krusval, pval = stats.kruskal(row, row2)
        df_kruskal.loc[i, j] = krusval
        df_pval.loc[i, j] = pval

df_kruskal

# df_pval
# df_k = df_kruskal.replace(0, np.nan) < 3.84146
x = 0
for i in df_k:
    x = x + df_k[i].values.sum()
x

df_pval.replace(0, np.nan) < 0.05

# Kruskal-Wallis on objectives

rbfd = {}
rbfn = []
for rbf in rbfs:
    name = rbf.__name__
    rbfn.append(name)
    rbfd[name] = pd.read_csv(f"refsets/{name}_refset.csv")

# difference in variance (rank sum)
# null hypothesis > no difference between groups

# https://www.youtube.com/watch?v=l86wEhUzkY4

# a = rbfd[rbfn[0]]['hydropower']
# b = rbfd[rbfn[1]]['hydropower']
# c = rbfd[rbfn[2]]['hydropower']
# d = rbfd[rbfn[3]]['hydropower']
# e = rbfd[rbfn[4]]['hydropower']
# f = rbfd[rbfn[5]]['hydropower']
# stats.kruskal(a,b,c,d,e,f)

a = [14, 1, 3, 9, 5, 7]
b = [10, 6, 8, 18, 11, 12]
c = [17, 16, 2, 13, 15, 4]
stats.kruskal(a, b, c)
# 2.858 is not greater than 5.99, reject the null hypothesis > there is a difference

# Hydropower

df_kruskal_hyd = pd.DataFrame(columns=rbfn)
df_pval_hyd = pd.DataFrame(columns=rbfn)
for i in rbfd:
    for j in rbfd:
        row1 = rbfd[i]["hydropower"]
        row2 = rbfd[j]["hydropower"]
        krusval, pval = stats.kruskal(row1, row2)
        df_kruskal_hyd.loc[i, j] = krusval
        df_pval_hyd.loc[i, j] = pval


df = df_kruskal_hyd.astype(float).round(2)
newcol = df.columns.to_list()
newcol = [x[:-4] for x in newcol]
df.columns = newcol
df.index = newcol

df_k = df_kruskal_hyd.replace(0, np.nan) < 3.84146
# x = 0
# for i in df_k:
#     x = x + df_k[i].values.sum()
# print(x)
df_k

df_pval_hyd.astype(float).round(2) < 0.05

# Atomic power plant

df_kruskal_atom = pd.DataFrame(columns=rbfn)
df_pval_atom = pd.DataFrame(columns=rbfn)
for i in rbfd:
    for j in rbfd:
        row1 = rbfd[i]["atomicpowerplant"]
        row2 = rbfd[j]["atomicpowerplant"]
        krusval, pval = stats.kruskal(row1, row2)
        df_kruskal_atom.loc[i, j] = krusval
        df_pval_atom.loc[i, j] = pval

df_k = df_kruskal_atom.replace(0, np.nan) < 3.84146
x = 0
for i in df_k:
    x = x + df_k[i].values.sum()
print(x)
df_k

df_pval_atom < 0.05

# Baltimore

df_kruskal_balt = pd.DataFrame(columns=rbfn)
df_pval_balt = pd.DataFrame(columns=rbfn)
for i in rbfd:
    for j in rbfd:
        row1 = rbfd[i]["baltimore"]
        row2 = rbfd[j]["baltimore"]
        krusval, pval = stats.kruskal(row1, row2)
        df_kruskal_balt.loc[i, j] = krusval
        df_pval_balt.loc[i, j] = pval

df_k = df_kruskal_balt.replace(0, np.nan) < 3.84146
x = 0
for i in df_k:
    x = x + df_k[i].values.sum()
print(x)
df_k

df_pval_balt < 0.05

# Chester

df_kruskal_ches = pd.DataFrame(columns=rbfn)
df_pval_ches = pd.DataFrame(columns=rbfn)
for i in rbfd:
    for j in rbfd:
        row1 = rbfd[i]["chester"]
        row2 = rbfd[j]["chester"]
        krusval, pval = stats.kruskal(row1, row2)
        df_kruskal_ches.loc[i, j] = krusval
        df_pval_ches.loc[i, j] = pval

df_k = df_kruskal_ches.replace(0, np.nan) < 3.84146
x = 0
for i in df_k:
    x = x + df_k[i].values.sum()
print(x)
df_k

df_pval_ches < 0.05

# Environment

df_kruskal_env = pd.DataFrame(columns=rbfn)
df_pval_env = pd.DataFrame(columns=rbfn)
for i in rbfd:
    for j in rbfd:
        row1 = rbfd[i]["environment"]
        row2 = rbfd[j]["environment"]
        krusval, pval = stats.kruskal(row1, row2)
        df_kruskal_env.loc[i, j] = krusval
        df_pval_env.loc[i, j] = pval

df_k = df_kruskal_env.replace(0, np.nan) < 3.84146
x = 0
for i in df_k:
    x = x + df_k[i].values.sum()
print(x)
df_k

df_pval_env < 0.05

# Recreation

df_kruskal_rec = pd.DataFrame(columns=rbfn)
df_pval_rec = pd.DataFrame(columns=rbfn)
for i in rbfd:
    for j in rbfd:
        row1 = rbfd[i]["recreation"]
        row2 = rbfd[j]["recreation"]
        krusval, pval = stats.kruskal(row1, row2)
        df_kruskal_rec.loc[i, j] = krusval
        df_pval_rec.loc[i, j] = pval

df_k = df_kruskal_rec.replace(0, np.nan) < 3.84146
x = 0
for i in df_k:
    x = x + df_k[i].values.sum()
print(x)
df_k

df_pval_rec < 0.05

# create index
df = df_kruskal_hyd.astype(float).round(2)
newcol = df.columns.to_list()
newcol = [x[:-4] for x in newcol]
df.columns = newcol
df.index = newcol

df = df_kruskal_rec.astype(float).round(2)
df2 = df_pval_rec.astype(float).round(2)
df.columns = newcol
df.index = newcol
df2.columns = newcol
df2.index = newcol
print("\\begin{table}[h]")
# print("\\parbox{.45\linewidth}{")
# print("\\centering")
print("\\begin{tabular}{c | " + " | ".join(["c"] * len(df.columns)) + "}")
# print(" & ".join([str(x) for x in newcol]) + " \\\\")
print(
    "& original & \makecell{squared\\\exponential} & \makecell{inverse\\\quadratic} & \makecell{inverse\\\multiquadric} & exponential & matern32 & matern52 \\\\"
)
print("\hline")
for i, row in df.iterrows():
    j = i.replace("_", " ")
    print(f"{j} & " + " & ".join([str(x) for x in row.values]) + " \\\\")
print("\\end{tabular}")
print("\\caption{Kruskal-Wallis test, recreation}")
print("\\end{table}")
print("")
# print("\\parbox{.45\linewidth}{")
# print("\\centering")
print("\\begin{table}[h]")
print("\\begin{tabular}{c | " + " | ".join(["c"] * len(df.columns)) + "}")
# print(" & " + " & ".join([str(x) for x in newcol]))
print(
    "& original & \makecell{squared\\\exponential} & \makecell{inverse\\\quadratic} & \makecell{inverse\\\multiquadric} & exponential & matern32 & matern52 \\\\"
)
print("\hline")
for i, row in df2.iterrows():
    j = i.replace("_", " ")
    print(f"{j} & " + " & ".join([str(x) for x in row.values]) + " \\\\")
print("\\end{tabular}")
print("\\caption{Kruskal-Wallis test, P value, recreation}")
# print("}")
print("\\end{table}")
